Move TSS pre-params marshalling code to dkg package
TSS pre-params are important only for DKG. Also, the dkg package is where the
pre-params pool lives.
